+++ test/application_system_test_case.rb
@@ -5,8 +5,12 @@ require "capybara/cuprite"
 class ApplicationSystemTestCase < ActionDispatch::SystemTestCase
   driven_by :cuprite, using: :chrome, screen_size: [ 1400, 900 ], options: {
     js_errors: true,
-    browser_options: { "no-sandbox": nil, "disable-gpu": nil },
-    process_timeout: 15,
-    timeout: 10
+    browser_options: {
+      "no-sandbox": nil,
+      "disable-gpu": nil,
+      "disable-dev-shm-usage": nil
+    },
+    process_timeout: 60,
+    timeout: 30
   }
 end